requirements.txt
Django>=1.10,<1.11
cfenv==0.5.2
dj-database-url==0.4.2
django-libsass==0.7
gunicorn==19.7.1
gevent==1.2.1
lxml==3.7.0
newrelic==2.82.0.62
psycopg2==2.7.1
requests==2.13.0
slacker==0.8.6
whitenoise==2.0.3
wagtail>=1.8,<1.9
bandit==1.4.0
# TODO: pin version
django-audit-log
Jinja2==2.9.6
django_jinja==2.3.1
CacheControl==0.11.5
cachetools==1.0.2

# Development
invoke==0.15.0
GitPython==1.0.1
requests-mock==1.3.0


# Data import
unicode-slugify==0.1.3
python-dateutil==2.6.0

# Content import
beautifulsoup4==4.5.1
django-storages==1.5.2
boto3==1.4.4

cg-django-uaa==1.0.1
